+++ packages/pluggableWidgets/badge-button-web/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+- We fixed an issue where custom button styles were not being applied correctly.
+
 ## [3.2.2] - 2026-02-09
 
 ### Added